requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements.txt requirements.in
#
aiohttp==1.0.5            # via discord.py
async-timeout==2.0.0      # via aiohttp
chardet==3.0.4            # via aiohttp
discord.py==0.16.12
multidict==4.1.0          # via aiohttp
pylast==2.1.0
six==1.11.0               # via pylast
websockets==3.4           # via discord.py
